refactor code to make it more OOP based and account for auto-allocation
