+++ src/main/java/duke/models/locker/Locker.java
@@ -1,8 +1,10 @@
-package duke.models;
+package duke.models.locker;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import duke.models.tag.Tag;
+
 
 import static java.util.Objects.requireNonNull;
