+++ src/main/java/duke/models/locker/SerialNumber.java
@@ -1,4 +1,4 @@
-package duke.models;
+package duke.models.locker;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonGetter;
@@ -11,6 +11,9 @@ public class SerialNumber {
 
     public static final String ERROR_MESSAGE = " Serial number can only be a positive integer with no more"
             + "than six digits and it cannot be empty";
+
+    public static final String CHECK_REGEX = "[0-9]+";
+
     public final String serialNumberForLocker;
 
     /**
@@ -27,8 +30,9 @@ public class SerialNumber {
         this.serialNumberForLocker = serialNumber;
     }
 
-    public static boolean checkIsValidSerialNumber(String test) {
-        return test.matches("[0-9]+") && test.length() <= 6;
+    public static boolean checkIsValidSerialNumber(String serialNumberForLocker) {
+        return serialNumberForLocker.matches(CHECK_REGEX)
+                && serialNumberForLocker.length() <= 6;
     }
 
     @JsonGetter("serialNumber")
