+++ src/main/java/duke/parser/AddBatchCommandParser.java
@@ -3,11 +3,11 @@ package duke.parser;
 import duke.exceptions.DukeException;
 import duke.logic.commands.AddBatchCommand;
 import duke.logic.commands.Command;
-import duke.models.Address;
-import duke.models.Locker;
-import duke.models.SerialNumber;
-import duke.models.Tag;
-import duke.models.Zone;
+import duke.models.locker.Address;
+import duke.models.locker.Locker;
+import duke.models.locker.SerialNumber;
+import duke.models.tag.Tag;
+import duke.models.locker.Zone;
 import duke.parser.utilities.MapTokensToArguments;
 import duke.parser.utilities.ParserTokenizer;
 import duke.parser.utilities.Token;
